Dividing into modules (3 squashed commits) wonder how branches are configured
separating code, reducing bullshit

travis for code, circle for emulator. happy builds, happy developer
