+++ .idea/findbugs-idea.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="FindBugs-IDEA">
+    <suppressWarningsClassName>java.lang.SuppressWarnings</suppressWarningsClassName>
+    <excludeFilterFiles>
+      <filter file="$PROJECT_DIR$/findbugs-filter.xml" enabled="true" />
+    </excludeFilterFiles>
+  </component>
+</project>
