+++ .travis.yml
@@ -13,25 +13,12 @@ android:
     - extra-android-m2repository
     - extra-android-support
     - extra-google-m2repository
-    - sys-img-armeabi-v7a-android-18
 
 licenses:
   - '.*'
 
-before_install:
-  - pip install --user codecov
-  
-before_script:
-  - echo no | android create avd --force -n test -t android-18 --abi armeabi-v7a
-  - emulator -avd test -no-skin -no-audio -no-window -force-32bit &
-  - android-wait-for-emulator
-  - adb shell input keyevent 82
-
 script:
-  - sh ci.sh
+  - ./gradlew check lint --stacktrace --console=plain -PdisablePreDex
 
-after_success:
-  - codecov
-  
 notifications:
   email: false
