+++ build.gradle
@@ -21,7 +21,7 @@ buildscript {
     dependencies {
         classpath 'com.android.tools.build:gradle:2.2.2'
         classpath 'com.neenbedankt.gradle.plugins:android-apt:1.8'
-        classpath 'me.tatarka:gradle-retrolambda:3.2.5'
+        classpath 'me.tatarka:gradle-retrolambda:3.3.1'
         classpath 'me.tatarka.retrolambda.projectlombok:lombok.ast:0.2.3.a2'
         classpath 'com.jakewharton.hugo:hugo-plugin:1.2.1'
         classpath 'com.fernandocejas.frodo:frodo-plugin:0.8.2'
@@ -62,7 +62,7 @@ ext {
     javaTargetCompatibility = JavaVersion.VERSION_1_8
 
     supportLibsVersion = '25.0.0'
-    daggerVersion = '2.2'
+    daggerVersion = '2.5'
     okHttpVersion = '3.2.0'
     retrofitVersion = '2.0.0-beta4'
     espressoVersion = '2.2.2'
@@ -79,8 +79,9 @@ ext.libraries.app = [
         javaxInject             : "javax.inject:javax.inject:1",
         javaxAnnotationApi      : "javax.annotation:javax.annotation-api:1.2",
 
-        rxJava                  : "io.reactivex:rxjava:1.1.0",
-        rxAndroid               : "io.reactivex:rxandroid:1.0.1",
+        rxJava                  : "io.reactivex:rxjava:1.2.2",
+        rxAndroid               : "io.reactivex:rxandroid:1.2.1",
+        rxAndroidProguard       : "com.artemzin.rxjava:proguard-rules:1.2.2.0",
         rxBinding               : 'com.jakewharton.rxbinding:rxbinding:0.3.0',
         rxBindingV7             : 'com.jakewharton.rxbinding:rxbinding-appcompat-v7:0.3.0',
 
@@ -93,6 +94,7 @@ ext.libraries.app = [
 
         autoValue               : "com.google.auto.value:auto-value:1.1",
         jodaTimeAndroid         : "net.danlew:android.joda:2.9.0",
+        jodaTime                : "joda-time:joda-time:2.9.5",
 
         supportAnnotations      : "com.android.support:support-annotations:$supportLibsVersion",
         supportAppCompat        : "com.android.support:appcompat-v7:$supportLibsVersion",
@@ -100,7 +102,8 @@ ext.libraries.app = [
         supportRecyclerView     : "com.android.support:recyclerview-v7:$supportLibsVersion",
         supportCardView         : "com.android.support:cardview-v7:$supportLibsVersion",
 
-        butterKnife             : "com.jakewharton:butterknife:7.0.1",
+        butterKnife             : "com.jakewharton:butterknife:8.4.0",
+        butterKnifeCompiler     : "com.jakewharton:butterknife-compiler:8.4.0",
         picasso                 : "com.squareup.picasso:picasso:2.5.2",
         timber                  : "com.jakewharton.timber:timber:4.1.0",
 
@@ -137,61 +140,45 @@ ext.libraries.test = [
 ]
 
 subprojects {
-    project.plugins.whenPluginAdded { plugin ->
-        if ('com.android.build.gradle.AppPlugin'.equals(plugin.class.name) || 'com.android.build.gradle.LibraryPlugin'.equals(plugin.class.name)) {
-            // enable or disable pre-dexing
-            project.android.dexOptions.preDexLibraries = rootProject.ext.preDexLibs
-        }
-    }
+    apply plugin: 'pmd'
 
-    plugins.apply('pmd')
-
-    task pmd(type: Pmd, group: 'verification') {
+    pmd {
         ignoreFailures = false // Fail early.
         ruleSets = ['java-basic', 'java-braces', 'java-strings', 'java-design', 'java-unusedcode']
-
-        source = fileTree('src/main/java')
     }
 
-    plugins.apply('findbugs')
+    apply plugin: 'findbugs'
 
-    task findbugs(type: FindBugs, group: 'verification') {
+    findbugs {
         ignoreFailures = false // Fail early.
         effort = 'default'
         reportLevel = 'low' // Report even low priority problems.
 
-        classes = files("${project.projectDir}/build/intermediates/classes/release")
-        source = fileTree('src/main/java')
-
-        // If somebody has an idea how to make this work with support libraries -> open a PR please.
-        classpath = files()
 
         excludeFilter = rootProject.file('findbugs-filter.xml')
+    }
 
+    tasks.withType(FindBugs) {
         reports {
             xml.enabled = false
             html.enabled = true
         }
     }
 
-    plugins.apply('checkstyle')
+    apply plugin: 'checkstyle'
 
-    task checkstyle(type: Checkstyle, group: 'verification') {
+    checkstyle {
         configFile rootProject.file('checkstyle.xml')
 
         ignoreFailures false // Fail early.
         showViolations true
-
-        source 'src'
-        include '**/*.java'
-
-        classpath = files()
     }
 
-    task checkstyleHtml(group: 'verification') << {
-        def path = "${project.buildDir}/reports/checkstyle/"
-        ant.xslt(in: path + 'checkstyle.xml',
-                style: rootProject.file("config/checkstyle-noframes-sorted.xsl"),
-                out: new File(path, 'checkstyle.html'))
+    tasks.withType(Checkstyle) {
+        reports {
+            xml.enabled true
+            html.enabled true
+            html.stylesheet resources.text.fromFile(rootProject.file('config/checkstyle-noframes-sorted.xsl'))
+        }
     }
 }
