+++ libservices/src/main/java/es/guillermoorellana/travisforandroid/api/entity/ApiBuildDetails.java
